% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create_schemas.R
\name{create_schemas}
\alias{create_schemas}
\title{creates database objects required for testing}
\usage{
create_schemas(conn)
}
\arguments{
\item{conn}{connection to the database opend using DBI::dbConnect()}
}
\description{
creates database objects required for testing
}

