% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/test_dbGetQuery.R
\name{test_dbGetQuery}
\alias{test_dbGetQuery}
\title{tests dbGetQuery() compatibility}
\usage{
test_dbGetQuery(conn)
}
\arguments{
\item{conn}{connection to the database opend using DBI::dbConnect()}
}
\value{
TRUE or FALSE indicating test pass/fail
}
\description{
tests dbGetQuery() compatibility
}

