% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/test_dbi_driver.R
\name{test_dbi_driver}
\alias{test_dbi_driver}
\title{tests a given dbi impelementation}
\usage{
test_dbi_driver(conn)
}
\arguments{
\item{conn}{connection to the database opend using DBI::dbConnect()}
}
\value{
list describing driver capabilities - see test_...() functions
  description
}
\description{
Given a handler to the database tests capabilities of the DBI driver.
}
\examples{
\dontrun{
 library(DBI)
 system('monetdbd start monetdb')
 handlers = list(
   SQLite = dbConnect(RSQLite::SQLite(), ":memory:"),
   MySQL = dbConnect(RMySQL::MySQL(), dbname = 'myDb'),
   PostgreSQL = dbConnect(RPostgreSQL::PostgreSQL(), dbname = 'myDb'),
   MonetDB = dbConnect(MonetDB.R::MonetDB.R(), 'pathToMyMonetDb')
 )
 sapply(handlers, test_dbi_driver)
}
}

