% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/test_dbListTables.R
\name{test_dbListTables}
\alias{test_dbListTables}
\title{tests behaviour of the dbListTables()}
\usage{
test_dbListTables(conn)
}
\arguments{
\item{conn}{connection to the database opend using DBI::dbConnect()}
}
\value{
numeric vector describing dbListTables() behaviour
}
\description{
As the dbListTables() specification in the DBI is simply wrong (if you want
to support schemas you can not return a character vector as a dbListTables()
result because it is impossible to correctly escape schema and table name
then), we can not provide a tests giving simple pass/fail results.

Instead of that we can only check out and describe some aspects of the
dbListTables() like if views are listed, if temporary tables are listed, etc.
}

