% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/test_dbReadTable.R
\name{test_dbReadTable}
\alias{test_dbReadTable}
\title{tests behaviour of the dbReadTable()}
\usage{
test_dbReadTable(conn)
}
\arguments{
\item{conn}{connection to the database opend using DBI::dbConnect()}
}
\value{
TRUE or FALSE indicating test pass/fail
}
\description{
tests behaviour of the dbReadTable()
}

