% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/test_dbReadTable_another_schema.R
\name{test_dbReadTable_another_schema}
\alias{test_dbReadTable_another_schema}
\title{tests behaviour of the dbReadTable() regarging tables in non-default
  schema}
\usage{
test_dbReadTable_another_schema(conn)
}
\arguments{
\item{conn}{connection to the database opend using DBI::dbConnect()}
}
\value{
logical vector describing dbReadTable() behaviour
}
\description{
Behaviour of the dbReadTable() regaring relations in schema other the default
one is not described. Due to that we can only describe the way the driver
behaves but we can not test if it meets the DBI interface or not.
}

