% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/test_dbSendQuery.R
\name{test_dbSendQuery}
\alias{test_dbSendQuery}
\title{tests dbSendQuery() compatibility}
\usage{
test_dbSendQuery(conn)
}
\arguments{
\item{conn}{connection to the database opend using DBI::dbConnect()}
}
\value{
TRUE or FALSE indicating test pass/fail
}
\description{
tests dbSendQuery() compatibility
}

