% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tidy_up.R
\name{tidy_up}
\alias{tidy_up}
\title{deletes all database objects created by create_schemas()}
\usage{
tidy_up(conn)
}
\arguments{
\item{conn}{connection to the database opend using DBI::dbConnect()}
}
\description{
deletes all database objects created by create_schemas()
}

